rename the module float::convert to float::conv to prepare for upcoming updates.
