+++ src/arm.rs
@@ -103,7 +103,7 @@ pub extern "C" fn __aeabi_uidiv(a: u32, b: u32) -> u32 {
 #[cfg(not(feature = "c"))]
 #[cfg_attr(not(test), no_mangle)]
 pub extern "C" fn __aeabi_ui2d(a: u32) -> f64 {
-    ::float::convert::__floatunsidf(a)
+    ::float::conv::__floatunsidf(a)
 }
 
 extern "C" {
